// app/components/types.ts
export interface UserAnswers {
  level: string;
  purpose: string[];
  skills: string[];
  partner: string;
  language: string;
}

export interface Option {
  name: string;
  emoji: string;
  color: string;
}

export type StepType =
  | "welcome"
  | "intro"
  | "level"
  | "purpose"
  | "skills"
  | "partner"
  | "recommendation";

export interface ScreenProps {
  onNext: () => void;
  onBack?: () => void;
  userAnswers?: UserAnswers;
  setUserAnswers?: (answers: UserAnswers) => void;
  name?: string;
  isSpeaking?: boolean;
}